src/third_party/blink/web_tests/virtual/web-components-v0-disabled/fast/dom/custom/document-register-namespace-expected.txt

CONSOLE ERROR: line 24: Uncaught TypeError: document.registerElement is not a function
Using document.registerElement() for extending HTML and non-HTML elements.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

FAIL successfullyParsed should be true. Was false.

TEST COMPLETE
